server/requirements.txt
fastapi
uvicorn
markitdown[pdf,docx,pptx]
lancedb
google-generativeai
python-multipart
python-dotenv
pydantic-settings
pyarrow
fastembed
